SFM: remove PBA frontend, use SparseBundleCPU directly
